Move all ffi symbols (success and error) into an ffi module
